+++ profiles.py
@@ -9,3 +9,14 @@ Lana = Character("Lana", "18", "Female", 0, """""", 0)
 Sid = Character("Sid", "16", "Male", 0, """""", 0)
 Connor = Character("Connor", "17", "Male", 0, """""", 0)
 Stephen = Character("Stephen", "16", "Male", 0, """""", 0)
+
+# Dialogue Locks
+# Lana:
+lana_dialoguecomplete1 = 0
+lana_dialoguecomplete2 = 0
+# Connor:
+connor_dialoguecomplete1 = 0
+connor_dialoguecomplete2 = 0
+connor_dialoguecomplete3 = 0
+# Stephen:
+stephen_dialoguecomplete1 = 0